Remove drawer folder on e2e and relocate all the files in their respective location
