--- e2e/pages/Drawer/TransactionHistoryView.js
@@ -1,3 +1,0 @@
-//import TestHelpers from '../helpers';
-
-export default class TransactionHistory {}
