+++ e2e/specs/permission-systems/permission-system-delete-wallet.spec.js
@@ -5,8 +5,8 @@ import OnboardingView from '../../pages/Onboarding/OnboardingView';
 import ProtectYourWalletView from '../../pages/Onboarding/ProtectYourWalletView';
 import CreatePasswordView from '../../pages/Onboarding/CreatePasswordView';
 import WalletView from '../../pages/WalletView';
-import Browser from '../../pages/Drawer/Browser';
-import SettingsView from '../../pages/Drawer/Settings/SettingsView';
+import Browser from '../../pages/Browser';
+import SettingsView from '../../pages/Settings/SettingsView';
 import TabBarComponent from '../../pages/TabBarComponent';
 import SkipAccountSecurityModal from '../../pages/modals/SkipAccountSecurityModal';
 import ConnectedAccountsModal from '../../pages/modals/ConnectedAccountsModal';
