+++ e2e/specs/permission-systems/permission-system-revoking-multiple-accounts.spec.js
@@ -1,7 +1,7 @@
 'use strict';
 import TestHelpers from '../../helpers';
 
-import Browser from '../../pages/Drawer/Browser';
+import Browser from '../../pages/Browser';
 
 import AccountListView from '../../pages/AccountListView';
 import TabBarComponent from '../../pages/TabBarComponent';
