+++ e2e/specs/quarantine/contract-nickname.failing.js
@@ -11,14 +11,14 @@ import WalletView from '../../pages/WalletView';
 import EnableAutomaticSecurityChecksView from '../../pages/EnableAutomaticSecurityChecksView';
 import LoginView from '../../pages/LoginView';
 
-import ContactsView from '../../pages/Drawer/Settings/Contacts/ContactsView';
-import SettingsView from '../../pages/Drawer/Settings/SettingsView';
+import ContactsView from '../../pages/Settings/Contacts/ContactsView';
+import SettingsView from '../../pages/Settings/SettingsView';
 
 import NetworkListModal from '../../pages/modals/NetworkListModal';
 import OnboardingWizardModal from '../../pages/modals/OnboardingWizardModal';
 import NetworkEducationModal from '../../pages/modals/NetworkEducationModal';
 import WhatsNewModal from '../../pages/modals/WhatsNewModal';
-import SecurityAndPrivacy from '../../pages/Drawer/Settings/SecurityAndPrivacy/SecurityAndPrivacyView';
+import SecurityAndPrivacy from '../../pages/Settings/SecurityAndPrivacy/SecurityAndPrivacyView';
 
 import TestHelpers from '../../helpers';
 import { acceptTermOfUse } from '../../viewHelper';
