+++ e2e/specs/quarantine/deeplinks.failing.js
@@ -6,13 +6,13 @@ import ConnectModal from '../../pages/modals/ConnectModal';
 import NetworkApprovalModal from '../../pages/modals/NetworkApprovalModal';
 import NetworkAddedModal from '../../pages/modals/NetworkAddedModal';
 
-import Browser from '../../pages/Drawer/Browser';
-import NetworkView from '../../pages/Drawer/Settings/NetworksView';
-import SettingsView from '../../pages/Drawer/Settings/SettingsView';
+import Browser from '../../pages/Browser';
+import NetworkView from '../../pages/Settings/NetworksView';
+import SettingsView from '../../pages/Settings/SettingsView';
 import LoginView from '../../pages/LoginView';
 import TransactionConfirmationView from '../../pages/TransactionConfirmView';
 
-import SecurityAndPrivacy from '../../pages/Drawer/Settings/SecurityAndPrivacy/SecurityAndPrivacyView';
+import SecurityAndPrivacy from '../../pages/Settings/SecurityAndPrivacy/SecurityAndPrivacyView';
 
 import WalletView from '../../pages/WalletView';
 import { importWalletWithRecoveryPhrase } from '../../viewHelper';
