+++ e2e/specs/settings/addressbook-tests.spec.js
@@ -1,9 +1,9 @@
 'use strict';
 import { SmokeCore } from '../../tags';
 import SendView from '../../pages/SendView';
-import SettingsView from '../../pages/Drawer/Settings/SettingsView';
-import ContactsView from '../../pages/Drawer/Settings/Contacts/ContactsView';
-import AddContactView from '../../pages/Drawer/Settings/Contacts/AddContactView';
+import SettingsView from '../../pages/Settings/SettingsView';
+import ContactsView from '../../pages/Settings/Contacts/ContactsView';
+import AddContactView from '../../pages/Settings/Contacts/AddContactView';
 import TabBarComponent from '../../pages/TabBarComponent';
 import WalletActionsModal from '../../pages/modals/WalletActionsModal';
 import AddAddressModal from '../../pages/modals/AddAddressModal';
