+++ e2e/specs/settings/delete-wallet.spec.js
@@ -3,9 +3,9 @@ import TestHelpers from '../../helpers';
 import { SmokeCore } from '../../tags';
 import OnboardingView from '../../pages/Onboarding/OnboardingView';
 import LoginView from '../../pages/LoginView';
-import SettingsView from '../../pages/Drawer/Settings/SettingsView';
-import SecurityAndPrivacyView from '../../pages/Drawer/Settings/SecurityAndPrivacy/SecurityAndPrivacyView';
-import ChangePasswordView from '../../pages/Drawer/Settings/SecurityAndPrivacy/ChangePasswordView';
+import SettingsView from '../../pages/Settings/SettingsView';
+import SecurityAndPrivacyView from '../../pages/Settings/SecurityAndPrivacy/SecurityAndPrivacyView';
+import ChangePasswordView from '../../pages/Settings/SecurityAndPrivacy/ChangePasswordView';
 import DeleteWalletModal from '../../pages/modals/DeleteWalletModal';
 import { loginToApp } from '../../viewHelper';
 import TabBarComponent from '../../pages/TabBarComponent';
