+++ e2e/viewHelper.js
@@ -5,11 +5,11 @@ import ImportWalletView from './pages/Onboarding/ImportWalletView';
 import MetaMetricsOptIn from './pages/Onboarding/MetaMetricsOptInView';
 import NetworkEducationModal from './pages/modals/NetworkEducationModal';
 import NetworkListModal from './pages/modals/NetworkListModal';
-import NetworkView from './pages/Drawer/Settings/NetworksView';
+import NetworkView from './pages/Settings/NetworksView';
 import OnboardingView from './pages/Onboarding/OnboardingView';
 import OnboardingCarouselView from './pages/Onboarding/OnboardingCarouselView';
 import OnboardingWizardModal from './pages/modals/OnboardingWizardModal';
-import SettingsView from './pages/Drawer/Settings/SettingsView';
+import SettingsView from './pages/Settings/SettingsView';
 import WalletView from './pages/WalletView';
 import WhatsNewModal from './pages/modals/WhatsNewModal';
 import Accounts from '../wdio/helpers/Accounts';
